feat: Auth 확인 로직을 컴포넌트로 변경
이전처럼 훅으로 로직 처리시 한 곳에서 일괄 처리가 불가능함
이를 layout에서 처리하기 위해서 컴포넌트로 구현

#9
